+++ python/ray/rllib/optimizers/__init__.py
@@ -1,0 +1,6 @@
+from ray.rllib.optimizers.async import AsyncOptimizer
+from ray.rllib.optimizers.local_sync import LocalSyncOptimizer
+from ray.rllib.optimizers.multi_gpu import LocalMultiGPUOptimizer
+
+
+__all__ = ["AsyncOptimizer", "LocalSyncOptimizer", "LocalMultiGPUOptimizer"]
