+++ python/ray/rllib/optimizers/multi_gpu.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from ray.rllib.optimizers.optimizer import Optimizer
+
+
+class LocalMultiGPUOptimizer(Optimizer):
+    pass  # TODO(ekl)
